Create Inactive Computers
Replace the value of $DaysInactive with the number of days you want to consider as the threshold for inactivity. The script uses the Get-ADComputer cmdlet to query Active Directory for computers with a LastLogonTimeStamp earlier than the calculated date. The result is then displayed, showing the computer names and their last logon timestamps.

Please note that the LastLogonTimeStamp attribute might not be completely accurate, as it is only updated periodically and might not reflect very recent logon activity. If you require more accurate information, you might need to use more complex methods involving event logs or other techniques.